requirements.txt
paho-mqtt
pyserial
Adafruit_Python_DHT